+++ resistify/tmbed/embed.py
@@ -18,73 +18,6 @@ import tempfile
 
 from pathlib import Path
 from transformers import logging
-from transformers import T5EncoderModel, T5Tokenizer
 
 
-class T5Encoder:
 
-    def __init__(self, model_path, use_gpu=True):
-        if not Path(model_path, 'config.json').exists():
-            self._download_models(model_path)
-
-        if use_gpu and torch.cuda.is_available():
-            self._load_models(model_path, torch.float16)
-            self.encoder_model = self.encoder_model.eval().cuda()
-        else:
-            self._load_models(model_path, torch.float32)
-            self.encoder_model = self.encoder_model.eval()
-
-        self.aa_map = str.maketrans('BJOUZ', 'XXXXX')
-
-    def _download_models(self, model_path):
-        model_name = 'Rostlab/prot_t5_xl_half_uniref50-enc'
-
-        Path(model_path).mkdir(parents=True, exist_ok=True)
-
-        with tempfile.TemporaryDirectory(dir=model_path) as temp_dir:
-            model_t = T5Tokenizer.from_pretrained(model_name,
-                                                  cache_dir=temp_dir,
-                                                  do_lower_case=False)
-
-            model_e = T5EncoderModel.from_pretrained(model_name,
-                                                     cache_dir=temp_dir,
-                                                     torch_dtype=torch.float16)
-
-            model_t.save_pretrained(model_path)
-            model_e.save_pretrained(model_path)
-
-        del model_t, model_e
-
-    def _load_models(self, model_path, dtype):
-        self.tokenizer = T5Tokenizer.from_pretrained(model_path,
-                                                     do_lower_case=False)
-
-        self.encoder_model = T5EncoderModel.from_pretrained(model_path,
-                                                            torch_dtype=dtype)
-
-    def device(self):
-        return self.encoder_model.device
-
-    def to_cpu(self):
-        self.encoder_model = self.encoder_model.cpu().float()
-
-    def to_cuda(self):
-        self.encoder_model = self.encoder_model.half().cuda()
-
-    def embed(self, sequences):
-        sequences = [s.upper().translate(self.aa_map) for s in sequences]
-
-        tokens = [' '.join(list(s)) for s in sequences]
-        tokens = self.tokenizer.batch_encode_plus(tokens,
-                                                  padding='longest',
-                                                  add_special_tokens=True)
-
-        device = self.encoder_model.device
-        input_ids = torch.tensor(tokens['input_ids'], device=device)
-        attention_mask = torch.tensor(tokens['attention_mask'], device=device)
-
-        with torch.no_grad():
-            embeddings = self.encoder_model(input_ids=input_ids,
-                                            attention_mask=attention_mask)
-
-        return embeddings.last_hidden_state.detach()
