+++ resistify/tmbed/utils.py
@@ -16,8 +16,7 @@
 import h5py
 import math
 import torch
-import numpy
-import random
+
 
 from torch.nn.utils.rnn import pad_sequence
 
@@ -27,127 +26,3 @@ from pathlib import Path
 from dataclasses import dataclass
 
 
-class OutFmt(str, Enum):
-    F0 = '0'
-    F1 = '1'
-    F2 = '2'
-    F3 = '3'
-    F4 = '4'
-
-
-@dataclass
-class ARGS:
-    fasta: Path = typer.Option(..., '--fasta', '-f',
-                               help='Input FASTA file.')
-    emb_in: Path = typer.Option(None, '--embeddings', '-e',
-                                help='Input embeddings file.')
-    emb_out: Path = typer.Option(..., '--embeddings', '-e',
-                                 help='Output embeddings file.')
-    predictions: Path = typer.Option(..., '--predictions', '-p',
-                                     help='Output predictions file.')
-    out_format: OutFmt = typer.Option('0', help='Prediction output format.')
-    batch_size: int = typer.Option(4000, help='Approximated batch size.')
-    use_gpu: bool = typer.Option(True, help='Use GPU if available.')
-    cpu_fallback: bool = typer.Option(True, help='Use CPU if GPU fails.')
-
-
-@dataclass
-class Protein:
-
-    header: str
-    sequence: str
-    length: int = 0
-    seq_hash: str = ''
-
-    def __post_init__(self):
-        self.length = len(self.sequence)
-        self.seq_hash = get_md5(self.sequence)
-
-
-def seed_all(seed):
-    random.seed(seed)
-    numpy.random.seed(seed)
-    torch.manual_seed(seed)
-
-    torch.backends.cudnn.benchmark = False
-    torch.backends.cudnn.deterministic = True
-
-
-def get_md5(string):
-    md5_hash = md5()
-
-    md5_hash.update(string.encode())
-
-    return md5_hash.hexdigest()
-
-
-def gaussian(x, std):
-    pi = torch.tensor(math.pi)
-    s2 = 2.0*torch.tensor(std).square()
-    x2 = torch.tensor(x).square().neg()
-
-    return torch.exp(x2 / s2) * torch.rsqrt(s2 * pi)
-
-
-def gaussian_kernel(kernel_size, std=1.0):
-    kernel = [gaussian(i - (kernel_size // 2), std)
-              for i in range(kernel_size)]
-
-    kernel = torch.tensor(kernel)
-    kernel = kernel / kernel.sum()
-
-    return kernel
-
-
-def make_batches(sequences, batch_size):
-    batches = []
-    num_prot = 1
-    last_idx = 0
-    max_size = 0
-    batch_size = abs(batch_size) ** 1.5
-
-    for idx, sequence in enumerate(sequences):
-        seq_size = len(sequence.sequence) ** 1.5
-        max_size = max(seq_size, max_size)
-
-        if (idx > 0) and (num_prot * max_size > batch_size):
-            batches.append((last_idx, idx))
-
-            num_prot = 1
-            last_idx = idx
-            max_size = seq_size
-
-        num_prot = num_prot + 1
-
-    batches.append((last_idx, idx + 1))
-
-    return batches
-
-
-def collate_batch(proteins, embeddings_file):
-    lengths = []
-    seq_hashes = []
-    embeddings = []
-
-    with h5py.File(embeddings_file, mode='r') as h5f:
-        for protein in proteins:
-            lengths.append(protein.length)
-            seq_hashes.append(protein.seq_hash)
-            embeddings.append(torch.from_numpy(h5f[protein.seq_hash][:]))
-
-    embeddings = pad_sequence(embeddings, batch_first=True, padding_value=0.0)
-
-    return seq_hashes, lengths, embeddings
-
-
-def make_mask(embeddings, lengths):
-    B, N, _ = embeddings.shape
-
-    mask = torch.zeros((B, N),
-                       dtype=embeddings.dtype,
-                       device=embeddings.device)
-
-    for idx, length in enumerate(lengths):
-        mask[idx, :length] = 1.0
-
-    return mask
